PORT: Use DWORD for thread-id (cleanup only)

+++ src/pl-thread.h
@@ -88,7 +88,7 @@ typedef struct _PL_thread_info_t
   pid_t		    pid;		/* for identifying */
 #endif
 #ifdef __WINDOWS__
-  unsigned long	    w32id;		/* Win32 thread HANDLE */
+  DWORD		    w32id;		/* Win32 thread HANDLE */
 #endif
   struct PL_local_data  *thread_data;	/* The thread-local data  */
   module_t	    module;		/* Module for starting goal */
